cleanup and re-org of files
