+++ Sources/Chart/MarkBuilder/AnyMark.swift
@@ -1,18 +1,12 @@
 //
-//  MarkBuilder.swift
+//  Mark.swift
 //
 //
-//  Created by Joseph Heck on 4/7/22.
+//  Created by Joseph Heck on 4/20/22.
 //
 
 import CoreGraphics
 import Foundation
-
-// MARK: - collection access to different kinds of 'mark' templates
-
-// reference for result builders:
-// https://docs.swift.org/swift-book/ReferenceManual/Attributes.html
-// https://docs.swift.org/swift-book/LanguageGuide/AdvancedOperators.html#ID630
 
 /// A type-erased Mark.
 public struct AnyMark: Mark {
@@ -34,28 +28,5 @@ public struct AnyMark: Mark {
 
     public func symbolsForMark(rangeLower: CGFloat, rangeHigher: CGFloat) -> [MarkSymbol] {
         wrappedSymbolsForMark(rangeLower, rangeHigher)
-    }
-}
-
-@resultBuilder
-public enum MarkBuilder {
-    // DEV NOTE: all of these builder expressions need to be `public static func` - miss
-    // the `public` part of it, and that excludes the function from being invoked during
-    // the compilation process, resulting in odd type errors (which I learned the hard way).
-
-    public static func buildExpression<T>(_ element: BarMark<T>) -> [AnyMark] {
-        [AnyMark(element)]
-    }
-
-    public static func buildExpression<T>(_ element: DotMark<T>) -> [AnyMark] {
-        [AnyMark(element)]
-    }
-
-    public static func buildExpression<T>(_ element: LineMark<T>) -> [AnyMark] {
-        [AnyMark(element)]
-    }
-
-    public static func buildBlock(_ components: [AnyMark]...) -> [AnyMark] {
-        Array(components.joined())
     }
 }
